Complete DeepSpeed integration for 85M parameter model
- Add DeepSpeed configuration with ZeRO Stage 2
- Scale model to 85M parameters with larger dimensions
- Fix data augmentation configuration compatibility
- Update trainer for DeepSpeed strategy integration
- Add memory optimization for 7.7GB GPU with CPU/SWAP offloading
- Fix loss function parameter matching
- Remove logs from git tracking

Model details:
- CAPE encoder: [64, 128, 256, 512] channels
- GNN: 256 hidden dim, 3 layers, 8 heads
- Transformer: 256 hidden dim, 6 layers, 16 heads
- Total: 85.4M parameters

Memory management:
- GPU: Active parameters only
- CPU: Optimizer states and gradients
- SWAP: Parameter overflow to 200GB NVME

+++ src/models/encoders/terrain_encoder.py
@@ -89,9 +89,9 @@ class TerrainEncoder(nn.Module):
     
         # Sobel filters for gradient computation
         sobel_x = torch.tensor([[-1, 0, 1], [-2, 0, 2], [-1, 0, 1]], 
-                              dtype=torch.float32, device=terrain_features.device)
+                      dtype=terrain_features.dtype, device=terrain_features.device)
         sobel_y = torch.tensor([[-1, -2, -1], [0, 0, 0], [1, 2, 1]], 
-                              dtype=torch.float32, device=terrain_features.device)
+                      dtype=terrain_features.dtype, device=terrain_features.device)
     
         # Reshape for grouped convolution - one filter per channel
         sobel_x = sobel_x.view(1, 1, 3, 3).repeat(num_channels, 1, 1, 1)
